requirements.txt
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.2
colorama==0.4.6
Django==5.2.5
django-environ==0.12.0
dnspython==2.7.0
environ==1.0
idna==3.10
iniconfig==2.1.0
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pymongo==4.3.3
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
flake8
black
djangorestframework
django-filter